refactor: split variants into modules
